requirements.txt
flask==2.3.3
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
scipy==1.14.1